feat: make markdown playground works
