+++ .dumi/theme/builtins/Playground/index.tsx
@@ -1,8 +1,6 @@
-import React from 'react';
-import { noop } from 'lodash-es';
-import SplitPane from 'react-split-pane';
-import { CodePreview } from '../../slots/CodePreview';
-import { CodeEditor } from '../../slots/CodeEditor';
+import React, { useContext } from 'react';
+import { ThemeAntVContext } from '../../context';
+import { CodeRunner } from '../../slots/CodeRunner';
 
 import styles from './index.module.less';
 
@@ -29,20 +27,15 @@ export type PlaygroundProps = {
  * Markdown 标签插件 Playground
  */
 const Playground: React.FC<PlaygroundProps> = ({ rid, path, ratio = 0.62, height = 400 }) => {
-  // @todo 逍为
-  /**
-   * 根据 context 传过来的所有案例，然后用 path 去匹配到对应的案例，拿到案例 demo 的信息，用于下面的渲染
-   */
+  /** 示例页面的元数据信息 */
+  const { meta }: any = useContext(ThemeAntVContext);
+  const { exampleTopics } = meta;
+  // '/case/area/demo/area5.ts'
+  const [_, topic, example, demo] = path.match(/\/([\w-]+)\/([\w-]+)\/demo\/([\w-]+)/i);
 
   return (
     <div className={styles.container} style={{ height }}>
-      {/** @ts-ignore */}
-      <SplitPane split="vertical" defaultSize={`${ratio * 100}%`} minSize={100}>
-        {/** @todo 逍为，和编辑器联动 */}
-        <CodePreview error={new Error('abc')} />
-        {/** @todo 逍为，获取源码内容和文件 */}
-        <CodeEditor source="" babeledSource="" relativePath={path} replaceId={rid} onError={noop} onFullscreen={noop} onDestroy={noop} onReady={noop} />
-      </SplitPane>
+      <CodeRunner exampleTopics={exampleTopics} topic={topic} example={example} demo={demo} />
     </div>
   );
 };
