+++ docs/manual/manual.md
@@ -3,3 +3,5 @@ title: manual
 order: 3
 ---
 manual
+
+<Playground path="/case/area/demo/area5.ts"></Playground>
